+++ apps/client/src/app/core/data/lazy-data.service.ts
@@ -242,7 +242,7 @@ export class LazyDataService {
         this.loaded$.next(true);
       });
 
-    const extractsPath = `/assets/extracts${environment.production ? '.' + extractsHash : ''}.json`;
+    const extractsPath = `/assets/extracts/extracts${environment.production ? '.' + extractsHash : ''}.json`;
 
     combineLatest([this.xivapi.getDCList(), this.getData<XivapiPatch[]>('https://xivapi.com/patchlist'), this.getData(extractsPath)]).subscribe(
       ([dcList, patches, extracts]) => {
@@ -306,7 +306,11 @@ export class LazyDataService {
     if (path.startsWith('http')) {
       url = path;
     } else {
-      url = `.${path}`;
+      if (this.platformService.isDesktop() || !environment.production || isPlatformServer(this.platform)) {
+        url = `.${path}`;
+      } else {
+        url = `https://cdn.ffxivteamcraft.com${path}`;
+      }
     }
     return this.http.get<T>(url);
   }