+++ apps/client/src/app/pages/simulator/components/macro-popup/macro-popup.component.ts
@@ -50,19 +50,19 @@ export class MacroPopupComponent implements OnInit {
 
   constructor(private l12n: LocalizedDataService, private i18n: I18nToolsService, private translator: TranslateService, public settings: SettingsService) {
     // TEMP: load settings and clean up localStorage
-    if (localStorage.getItem("macros:addecho") !== null){
-      this.addEcho = this.settings.macroEcho = localStorage.getItem("macros:addecho") !== 'false';
-      localStorage.removeItem("macros:addecho");
+    if (localStorage.getItem('macros:addecho') !== null) {
+      this.addEcho = this.settings.macroEcho = localStorage.getItem('macros:addecho') !== 'false';
+      localStorage.removeItem('macros:addecho');
     }
 
-    if (localStorage.getItem("macros:macrolock") !== null){
+    if (localStorage.getItem('macros:macrolock') !== null) {
       this.macroLock = this.settings.macroLock = localStorage.getItem('macros:macrolock') === 'true';
-      localStorage.removeItem("macros:macrolock");
+      localStorage.removeItem('macros:macrolock');
     }
 
-    if (localStorage.getItem("macros:consumables") !== null){
+    if (localStorage.getItem('macros:consumables') !== null) {
       this.addConsumables = this.settings.macroConsumables = localStorage.getItem('macros:consumables') === 'true';
-      localStorage.removeItem("macros:consumables");
+      localStorage.removeItem('macros:consumables');
     }
   }
 
@@ -77,6 +77,9 @@ export class MacroPopupComponent implements OnInit {
     this.settings.macroCompletionMessage = this.macroCompletionMessage;
 
     this.macro = this.macroLock ? [['/mlock']] : [[]];
+    if (this.addConsumables) {
+      this.macro[0].push(this.getConsumablesNotification());
+    }
     this.totalDuration = 0;
     let totalLength = 0;
     this.rotation.forEach((action, actionIndex) => {
@@ -101,7 +104,7 @@ export class MacroPopupComponent implements OnInit {
       }
 
       let doneWithChunk: boolean;
-      if (this.breakBeforeByregotsBlessing && actionIndex < this.rotation.length - 1 && this.rotation[actionIndex+1].is(ByregotsBlessing)) {
+      if (this.breakBeforeByregotsBlessing && actionIndex < this.rotation.length - 1 && this.rotation[actionIndex + 1].is(ByregotsBlessing)) {
         doneWithChunk = true;
       } else if (macroFragment.length === 14 && this.addEcho && this.rotation.length > totalLength + 1) {
         doneWithChunk = true;