+++ apps/client/src/app/pages/simulator/simulator.module.ts
@@ -50,6 +50,8 @@ import { DragDropModule } from '@angular/cdk/drag-drop';
 import { SimulationResultComponent } from './components/simulation-result/simulation-result.component';
 import { AntdSharedModule } from '../../core/antd-shared.module';
 import { NzPaginationModule } from 'ng-zorro-antd/pagination';
+import { NzCheckboxModule } from 'ng-zorro-antd/checkbox';
+import { LazyScrollModule } from '../../modules/lazy-scroll/lazy-scroll.module';
 
 const routes: Routes = [
   {
@@ -137,7 +139,9 @@ const routes: Routes = [
     SettingsModule,
     DirtyModule,
     DragDropModule,
-    NzPaginationModule
+    NzPaginationModule,
+    NzCheckboxModule,
+    LazyScrollModule
   ],
   declarations: [
     CustomSimulatorPageComponent,